requirements.txt
selenium
webdriver-manager
python-dotenv
yagmail
twilio
pandas
